finish token expire interface
added the hook to every page and the nav so that no matter what page the user is on, if their token expires they're prompted to login and the nav resets to the signed out state. there's probably a more efficient way to do this but this was the method i could get working quickly.